Handle icore errors when fetching game url.
As in Voidbridge, we need to handle the case when iCore
returns an error when we try to fetch the game URL using
the external game id. ProviderTrait maps the iCore error
to some message but this message needs to be passed to
the JS code that made the request.